Routes cleanup
