+++ sites/partners/layouts/application.tsx
@@ -23,14 +23,8 @@ const Layout = (props) => (
         notice=""
         title={t("nav.siteTitle")}
       >
-        <LocalizedLink href="/properties" className="navbar-item">
-          {t("nav.properties")}
-        </LocalizedLink>
-        <LocalizedLink href="/listings" className="navbar-item">
+        <LocalizedLink href="/" className="navbar-item">
           {t("nav.listings")}
-        </LocalizedLink>
-        <LocalizedLink href="/applications" className="navbar-item">
-          {t("nav.applications")}
         </LocalizedLink>
       </SiteHeader>
       <main>{props.children}</main>
